pyproject.toml
[project]
name = "contextual-rag-local"
version = "0.1.0"
description = "A fully local, completely free RAG system with contextual retrieval"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Sergio", email = "user@example.com"}
]

dependencies = [
    "streamlit>=1.28.0",
    "chromadb>=0.4.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pypdf>=3.17.0",
    "PyPDF2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.0.290",
]
